src/main/java/daily/JumpGame.java
package daily;

public class JumpGame {
	/**
	 * 
	 * @param nums - You are given an integer array nums. You are initially
	 *             positioned at the array's first index, and each element in the
	 *             array represents your maximum jump length at that position.
	 * @return true if you can reach the last index, or false otherwise.
	 */

	public boolean canJump(int[] nums) {
		int lastPos = nums.length - 1;
		for (int i = nums.length - 1; i >= 0; i--) {
			if (i + nums[i] >= lastPos) {
				lastPos = i;
			}
		}
		return lastPos == 0;
	}

}
